+++ src/ui/icons.tsx
@@ -428,3 +428,11 @@ export function PlugIcon(props: IconProps) {
     </svg>
   );
 }
+
+export function BookmarkIcon(props: IconProps) {
+  return (
+    <svg {...svgDefaults(props)}>
+      <path d="M19 21l-7-5-7 5V5a2 2 0 0 1 2-2h10a2 2 0 0 1 2 2z" />
+    </svg>
+  );
+}